+++ scripts/powerUps.js
@@ -33,8 +33,6 @@ class Powerup {
     const powerup = Powerup.ordemPowerups[index];
     powerup.comprado = true;
 
-    Powerup.powerupsComprados.push(powerup);
-    Powerup.ordemPowerups.splice(index, 1);
     this.atualizarFront();
     esconderPowerupDetalhado();
   }
@@ -42,7 +40,9 @@ class Powerup {
   static atualizarFront() {
     filaPowerups.innerHTML = "";
 
-    if (Powerup.ordemPowerups.length === 0) {
+    const restantes = Powerup.ordemPowerups.filter(pwp => pwp.comprado === false)
+
+    if (restantes.length === 0) {
       filaPowerups.innerHTML = `
         <div class="itemPowerup">Parabéns! Você comprou todos</div>
       `;
@@ -50,19 +50,20 @@ class Powerup {
     }
 
     for (let i = 0; i < 3; i++) {
-      if (Powerup.ordemPowerups.length <= i) {
+      if (restantes.length <= i) {
         filaPowerups.innerHTML += `<div class="itemPowerup"></div>`;
         continue;
       }
 
-      const powerup = Powerup.ordemPowerups[i];
+      const powerup = restantes[i];
+      const clickAction = `Powerup.ordemPowerups[${powerup.index}].comprarPowerup(${powerup.index})`
 
       filaPowerups.innerHTML += `
         <div 
           class="itemPowerup"
           title="${powerup.descricao}"
-          onclick="Powerup.ordemPowerups[${i}].comprarPowerup(${i})"
-          onmouseenter="mostrarPowerupDetalhado(${i})"  
+          onclick="${clickAction}"
+          onmouseenter="mostrarPowerupDetalhado(${powerup.index})"  
           onmouseleave="esconderPowerupDetalhado()"  
         >
           ${powerup.nome}
@@ -75,25 +76,25 @@ class Powerup {
 const infoPowerups = {
   pw1: new Powerup(
     "Cliques também",
-    `+1 cpc para cada 10 ${upgrade1.nome}`,
+    `+1 bpc para cada 10 ${upgrade1.nome}`,
     250,
-    `upgrade1.cpc += 0.1; poderClique += upgrade1.quantidade*upgrade1.cpc`,
+    `upgrade1.bpc += 0.1; poderClique += upgrade1.quantidade*upgrade1.bpc`,
   ),
   pw2: new Powerup(
     "Melhores preços",
-    `taxa de aumento 5% menor para ${upgrade3.nome}`,
+    `taxa de aumento de preço 5% menor para ${upgrade3.nome}`,
     700,
     `upgrade3.taxaPreco *= 0.95;`,
   ),
-  pw3: new Powerup(
-    "No futuro fará sentido",
-    `+5 ${upgrade2.nome}`,
-    1500,
-    `upgrade2.quantidade +=5;
-     poderClique+= upgrade2.cpc*5;
-     document.getElementsByClassName("quantidadeUpgrade")[3].innerText = upgrade2.quantidade
-    `,
-  ),
+  // pw3: new Powerup(
+  //   "No futuro fará sentido",
+  //   `+5 ${upgrade2.nome}`,
+  //   1500,
+  //   `upgrade2.quantidade +=5;
+  //    poderClique+= upgrade2.bpc*5;
+  //    document.getElementsByClassName("quantidadeUpgrade")[1].innerText = upgrade2.quantidade
+  //   `,
+  // ),
   pw4: new Powerup(
     "Cultivo rápido",
     `${upgrade1.nome} 10% mais rápido`,
@@ -104,7 +105,7 @@ const infoPowerups = {
     "Pratos maiores",
     `+3 batatas/segundo na ${upgrade1.nome}`,
     4900,
-    `upgrade1.cps += 3; batatasPS += upgrade1.quantidade * 3`,
+    `upgrade1.bpi += 3; batatasPS += upgrade1.quantidade * 3`,
   ),
 
 
@@ -113,7 +114,7 @@ const infoPowerups = {
     "Cliques doces",
     `+1 batata/clique para cada 4 ${upgrade2.nome} `,
     8000,
-    `upgrade2.cpc += 0.25; poderClique += upgrade2.quantidade*upgrade2.cpc`,
+    `upgrade2.bpc += 0.25; poderClique += upgrade2.quantidade*upgrade2.bpc`,
   ),
 
   pw6: new Powerup(
@@ -126,26 +127,26 @@ const infoPowerups = {
   pw8: new Powerup(
     "Enxada afiada",
     `+3 batata/clique na ${upgrade5.nome}`,
-    50,
-    `upgrade5.cpc += 3; poderClique += upgrade5.quantidade * 3`,
+    13500,
+    `upgrade5.bpc += 3; poderClique += upgrade5.quantidade * 3`,
   ),
   pw9: new Powerup(
     "Limão, sal, azeite...",
     `+5 batatas/segundo na ${upgrade1.nome}`,
-    50,
-    `upgrade1.cps += 5; batatasPS += upgrade1.quantidade * 5`,
+    20000,
+    `upgrade1.bpi += 5; batatasPS += upgrade1.quantidade * 5`,
   ),
   pw10: new Powerup(
     "Crocância e sabor",
     `+10 batatas/segundo na ${upgrade2.nome}`,
-    50,
-    `upgrade2.cps += 10; batatasPS += upgrade2.quantidade * 10`,
+    33000,
+    `upgrade2.bpi += 10; batatasPS += upgrade2.quantidade * 10`,
   ),
   pw11: new Powerup(
     "Adubo",
     `6 batatas/clique na ${upgrade5.nome}`,
-    50,
-    `upgrade5.cpc += 6; poderClique += upgrade5.quantidade * 6`,
+    37000,
+    `upgrade5.bpc += 6; poderClique += upgrade5.quantidade * 6`,
   ),
 };
 
@@ -163,7 +164,7 @@ function mostrarPowerupDetalhado(indexPowerup) {
   const descricao = document.getElementById("descricaoPowerup");
 
   nome.textContent = powerup.nome;
-  preco.textContent = powerup.preco + " batatas";
+  preco.textContent = transformNum(powerup.preco, 2, true) + " batatas";
   descricao.textContent = powerup.descricao;
 
   document.getElementById("infoPowerups").style.display = "block";