+++ scripts/salvar.js
@@ -3,6 +3,20 @@ function salvarTemporario() {
   document.querySelector("#popupSalvar").style.display = "block";
   setTimeout(() => (document.querySelector("#popupSalvar").style.display = "none"), 2000);
 
+  const filteredUpgrades = Upgrades.upgradesExistentes.map(up =>{
+    return up
+    // return [up.quantidade || 0, up.preco || NaN]
+  })
+
+  const filteredConquistas = Conquistas.conquistasTotais
+    .filter(conq => conq && conq.completa == true)
+    .map(conq => {return conq.index})
+
+  const filteredPowerups = Powerup.ordemPowerups
+  .filter(pwp => pwp.comprado === true)
+  .map(pwp =>{return pwp.index})
+
+  
   const save = {
     geral: {
       cliques: cliques,
@@ -17,12 +31,9 @@ function salvarTemporario() {
       BatataPS: batatasPS,
       poderClique: poderClique,
     },
-    conquistas: Conquistas.conquistasLiberadas,
+    conquistas: filteredConquistas,
     upgrades: Upgrades.upgradesExistentes,
-    powerups: {
-      pendentes: Powerup.ordemPowerups,
-      desbloqueados: Powerup.powerupsComprados
-    }
+    powerups: filteredPowerups
   }
 
   localStorage.setItem("CCconfig", JSON.stringify(save))
@@ -34,23 +45,25 @@ function carregarSave(){
     const save = JSON.parse(localStorage.getItem("CCconfig"))
 
     // Configurações gerais
-    cliques = save.geral.cliques
-    Upgrades.numeroDeUpgrades = save.geral.numeroDeUpgrades
-    SFXligado = save.geral.SFXligado
-    volumeSFX.value = save.geral.volume
-    seg = save.geral.seg
+    cliques = save.geral.cliques ?? 0
+    Upgrades.numeroDeUpgrades = save.geral.numeroDeUpgrades ?? 0
+    SFXligado = save.geral.SFXligado ?? true
+    volumeSFX.value = save.geral.volume ?? "0.5"
+    seg = save.geral.seg ?? 0
 
     // Configurações sobre batatas
-    batatas = save.batatas.bancoBatata
-    batataTotal = save.batatas.batataTotal
-    batatasPS = save.batatas.BatataPS
-    poderClique = save.batatas.poderClique
+    batatas = save.batatas.bancoBatata ?? 0
+    batataTotal = save.batatas.batataTotal ?? 0
+    batatasPS = save.batatas.BatataPS ?? 0
+    poderClique = save.batatas.poderClique ?? 0
 
     // Configurações das conquistas
-    Conquistas.numeroConquistasLiberadas = save.conquistas.length
-    Conquistas.conquistasLiberadas = save.conquistas
-    Conquistas.conquistasLiberadas.forEach(c =>{
-      Conquistas.conquistasTotais[c.index].completa = true
+    save.conquistas.forEach(index =>{
+      if(index === null || index === undefined){
+        console.log("Index não declarado")
+        return
+      }
+      Conquistas.conquistasTotais[index].completa = true
     })
 
     // Configurações dos upgrades
@@ -59,42 +72,17 @@ function carregarSave(){
       if(!upgd){
         return
       }
-      upgd.quantidade = configUpgrades[i].quantidade
-      upgd.preco = configUpgrades[i].preco
+      upgd.quantidade = configUpgrades[i].quantidade ?? upgd.quantidade
+      upgd.preco = configUpgrades[i].preco ?? upgd.preco
+      upgd.bpi = configUpgrades[i].bpi ?? upgd.bpi
+      upgd.bpc = configUpgrades[i].bpc ?? upgd.bpc
     })
 
-    if(save.powerups.desbloqueados.length !== 0){
-      const backupPendentes = structuredClone(Powerup.ordemPowerups)
-
-      try {
-
-        Powerup.ordemPowerups = []
-        Powerup.powerupsComprados = []
+    if(save.powerups.length !== 0){
+      save.powerups.forEach(index =>{
+        Powerup.ordemPowerups[index].comprado = true
         Powerup.atualizarFront()
-
-        // Esse código não merece ser comentado, tá horrível e espero q vc nunca descubra oq eu fiz
-        save.powerups.desbloqueados.forEach(powerup =>{
-          const {nome, descricao, preco, efeito, cumulativo} = {...powerup}
-          const novoPowerup = new Powerup(nome, descricao, preco, efeito)
-          novoPowerup.comprado = true;
-          Powerup.powerupsComprados.push(novoPowerup)
-          Powerup.ordemPowerups.pop()
-        })
-
-        
-        save.powerups.pendentes.forEach(powerup =>{
-          const {nome, descricao, preco, efeito, cumulativo} = {...powerup}
-          const novoPowerup = new Powerup(nome, descricao, preco, efeito)
-        })
-
-        Powerup.atualizarFront()
-      } catch (error) {
-        console.log(error)
-        Powerup.ordemPowerups = backupPendentes
-        
-      }
-
+      })
     }
-
   }
 }